chore: base rust client
