--- experimental/client-rust/.bumpversion.cargo.cfg
@@ -1,37 +1,0 @@
-[bumpversion]
-current_version = 0.17.0-beta.9
-commit = False
-tag = False
-allow_dirty = True
-parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(-(?P<prekind>[a-z]+)\.(?P<pre>\d+))?(\.(?P<devkind>[a-z]+)\.(?P<dev>\d+))?
-serialize =
-	{major}.{minor}.{patch}-{prekind}.{pre}.{devkind}.{dev}
-	{major}.{minor}.{patch}.{devkind}.{dev}
-	{major}.{minor}.{patch}-{prekind}.{pre}
-	{major}.{minor}.{patch}
-
-[bumpversion:part:pre]
-first_value = 1
-
-[bumpversion:part:prekind]
-optional_value = _
-values =
-	_
-	alpha
-	beta
-	rc
-	_
-
-[bumpversion:part:dev]
-first_value = 1
-
-[bumpversion:part:devkind]
-optional_value = _
-values =
-	_
-	dev
-	_
-
-[bumpversion:file:client/Cargo.toml]
-search = version = "{current_version}"
-replace = version = "{new_version}"
