--- experimental/client-rust/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-# name = "wandb"
-name = "wandbrs"           # 🪄🍻
-version = "0.19.0-alpha.1"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-crate-type = ["lib", "cdylib"]
-
-# [[bin]]
-# name = "binder"
-# path = "src/binder.rs"
-
-[dependencies]
-bytes = "1.5.0"
-byteorder = "1.5.0"
-prost = "0.13"
-prost-types = "0.13"
-rand = "0.8.5"
-pyo3 = { version = "0.20", features = ["abi3-py37", "extension-module"] }
-tempfile = "3.10.1"
-prost-build = "0.13"
-uuid = "1.10.0"
-fork = "0.2"
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-sentry = "0.34.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-colored = "2.1.0"
-indicatif = "0.17.8"
-chrono = "0.4.38"
-numpy = "0.20.0"
-image = "0.24.9"
-sha2 = "0.10.8"
-
-[build-dependencies]
-pyo3-build-config = "0.22"
-prost-build = "0.13"
-tempfile = "3.10.1"
