--- experimental/client-rust/pyproject.toml
@@ -1,12 +1,0 @@
-[build-system]
-requires = ["maturin>=1.0,<2.0"]
-build-backend = "maturin"
-
-[tool.maturin]
-# Build artifacts with the specified Cargo profile
-profile = "release"
-features = ["pyo3/extension-module"]
-include = [
-    { path = "wandbrs/wandb-core", format = "wheel" },
-    { path = "wandbrs/gpu_stats", format = "wheel" },
-]
