--- experimental/client-rust/src/lib.rs
@@ -1,82 +1,0 @@
-use pyo3::prelude::*;
-use sentry;
-use std::sync::OnceLock;
-use tracing;
-use tracing_subscriber;
-
-pub mod connection;
-pub mod launcher;
-pub mod printer;
-pub mod run;
-pub mod session;
-pub mod settings;
-pub mod wandb_internal;
-
-fn parse_bool(s: &str) -> bool {
-    match s.to_lowercase().as_str() {
-        "true" | "1" => true,
-        "false" | "0" => false,
-        _ => true,
-    }
-}
-
-/// Communication layer between user code and nexus
-
-pub fn get_core_version() -> &'static str {
-    static VERSION: OnceLock<String> = OnceLock::new();
-    VERSION.get_or_init(|| {
-        let version = env!("CARGO_PKG_VERSION");
-        version
-            .replace("-alpha.", "a")
-            .replace("-beta.", "b")
-            .replace("-rc.", "rc")
-    })
-}
-
-
-#[pyfunction]
-pub fn init(project: Option<String>, settings: Option<settings::Settings>) -> PyResult<run::Run> {
-    let mut settings = settings.unwrap_or_default();
-
-    if let Some(project) = project {
-        settings.set_project(project);
-    }
-
-    let sess = session::Session::new(settings)?;
-    sess.init_run(None)
-}
-
-/// A Python module implemented in Rust. The name of this function must match
-/// the `lib.name` setting in the `Cargo.toml`, else Python will not be able to
-/// import the module.
-#[pymodule]
-fn wandbrs(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
-    let error_reporting_enabled = std::env::var("WANDB_ERROR_REPORTING")
-        .map(|v| parse_bool(&v))
-        .unwrap_or(true);
-
-    let dsn: Option<sentry::types::Dsn> = if error_reporting_enabled {
-        "https://9e9d0694aa7ccd41aeb5bc34aadd716a@o151352.ingest.us.sentry.io/4506068829470720"
-            .parse()
-            .ok()
-    } else {
-        None
-    };
-
-    let _guard = sentry::init(sentry::ClientOptions {
-        dsn,
-        release: sentry::release_name!(),
-        ..Default::default()
-    });
-
-    let log_level = tracing::Level::INFO;
-    // let log_level = tracing::Level::DEBUG;
-    tracing_subscriber::fmt().with_max_level(log_level).init();
-
-    m.add("__version__", get_core_version())?;
-    m.add_function(wrap_pyfunction!(init, m)?)?;
-    m.add_class::<settings::Settings>()?;
-    m.add_class::<session::Session>()?;
-    m.add_class::<run::Run>()?;
-    Ok(())
-}
