+++ experimental/rust-sdk/.gitignore
@@ -1,5 +1,9 @@
 # Generated by Rust toolchain.
 target/*
+examples/basic/target/*
 
 # Generated from proto files by build process.
 src/wandb_internal.rs
+
+# Generated binary files.
+bin/*
