+++ experimental/rust-sdk/Cargo.toml
@@ -11,3 +11,24 @@ readme = "README.md"
 
 [lib]
 path = "src/lib.rs"
+
+[dependencies]
+sentry = "0.34.0"
+tracing = "0.1.40"
+tracing-subscriber = "0.3.18"
+prost = "0.13"
+prost-types = "0.13"
+rand = "0.8.5"
+byteorder = "1.5.0"
+indicatif = "0.17.8"
+colored = "2.1.0"
+fork = "0.2"
+chrono = "0.4.38"
+tempfile = "3.10.1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+
+[build-dependencies]
+tempfile = "3.10.1"
+prost-build = "0.13"
