+++ experimental/rust-sdk/examples/basic/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "basic"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+wandb = { path = "../../." }
